version 1.0.0 release

Closes FRINGE-174, FRINGE-211, and FRINGE-210

See merge request arcticwind/seafringe/teresa!32

+++ teresa/coregistration.py
@@ -10,6 +10,7 @@ from . import graphs
 from . import processor
 from .log import LOG_FNAME
 from datetime import datetime
+from itertools import product
 
 # https://stackoverflow.com/questions/46641078/how-to-avoid-circular-dependency-caused-by-type-hinting-of-pointer-attributes-in
 import typing
@@ -30,32 +31,65 @@ class CoregistrationError(RuntimeError):
 
 
 class Coregistration(abc.ABC):
-    @abc.abstractmethod
-    def coregister(self):
-        ...
 
-
-class Sentinel1Coregistration(Coregistration):
-    # concrete creater
     def __init__(
         self,
         slc_pair: stack.SlcPair,
         output_dir: str,
         polarization: str = "vv",
         dry_run: bool = True,
+        prune: bool = True,
+        radarcode_dem: bool = False,  # by default don't radarcode dem
         coreg_processor: processor.GptProcessor = processor.GptProcessor(),
     ):
-        self.graph = "None"
+        # concrete implementation of init function
         self.slc_pair = slc_pair
         self.output_dir = output_dir
         self.polarization = polarization
         self.dry_run = dry_run
+        self.prune = prune
+        self.radarcode_dem = radarcode_dem
         self._processor = coreg_processor  # interface
+
+    @abc.abstractmethod
+    def coregister(self):
+        ...
+
+    def _radarcode_dem(self):
+        """concrete implementation of geocoding DEM using the
+        interferogram function in snap. The implementation of geocoding DEM
+        should be the same regardless of the sensors. This is why we put this
+        function into the parent class.
+        """
+
+        input_pair = os.path.join(self.output_dir, COREG_DIR, self.slc_pair.master.date, "merged.dim")
+        output_path = os.path.join(self.output_dir, COREG_DIR, "DEM")
+
+        # Execute the actual coregistration
+        self._processor.process(
+            graph=graphs.GptGraph.radarcode_dem(),
+            input_pair=input_pair,
+            output_path=output_path,
+            dry_run=self.dry_run,
+        )
+
+        if self.dry_run:
+            master_datestr = format_date(self.slc_pair.master.date)
+            logger.debug("DRY RUN: Creating dummy 'DEM' folder for testing purpose.")
+            os.makedirs(os.path.join(output_path, "merged") + ".data", exist_ok=True)
+            open(os.path.join(output_path, "merged") + ".dim", "w").close()
+            open(os.path.join(output_path, "merged.data", "elevation.hdr"), "w").close()
+            open(os.path.join(output_path, "merged.data", f"coh_VV_{master_datestr}_{master_datestr}.hdr"), "w").close()
+
+
+class Sentinel1Coregistration(Coregistration):
 
     def coregister(self):
         self._prepare()
         self._coregister()
         self._merge()
+        if self.radarcode_dem:  # if radarcode dem, do it before the prune
+            self._radarcode_dem()
         self._finalize()
 
     def _prepare(self):
@@ -82,12 +116,13 @@ class Sentinel1Coregistration(Coregistration):
             f"iw{nsubswath}",
         )
 
-        logger.info(  # logging before exeuution
-            "COREGISTERING master %s and slave %s for swath IW%s:",
-            self.slc_pair.master.date,
-            self.slc_pair.slave.date,
-            nsubswath,
-        )
+        if not self.dry_run:
+            logger.info(  # logging before exeuution
+                "COREGISTERING master %s and slave %s for swath IW%s:",
+                self.slc_pair.master.date,
+                self.slc_pair.slave.date,
+                nsubswath,
+            )
 
         # Execute the actual coregistration
         self._processor.process(
@@ -101,11 +136,12 @@ class Sentinel1Coregistration(Coregistration):
         )
 
         self.slc_pair.slave.append(destination=output_path)
-        logger.info(f"COREGISTERING slave {self.slc_pair.slave.date} completed.")
+        if not self.dry_run:
+            logger.info(f"COREGISTERING slave {self.slc_pair.slave.date} swath IW{nsubswath} completed.")
 
         if self.dry_run:
-            logger.debug("DRY-RUN: Creating dummy folders/files for testing purpose.")
-            for i in range(1, 4):
+            logger.debug("DRY RUN: Creating dummy folders/files for testing purpose.")
+            for _ in range(1, 4):
                 os.makedirs(output_path + ".data", exist_ok=True)
                 open(output_path + ".dim", "w").close()
 
@@ -122,7 +158,8 @@ class Sentinel1Coregistration(Coregistration):
             "merged",  # save in merged directory
         )
 
-        logger.info(f"MERGING subswaths of image {self.slc_pair.slave.date}:")
+        if not self.dry_run:
+            logger.info(f"MERGING subswaths of image {self.slc_pair.slave.date}:")
 
         input_subswaths = {
             f"input_subswath{i+1}": path + ".dim"  # merge takes in .dim file
@@ -133,26 +170,28 @@ class Sentinel1Coregistration(Coregistration):
             graph, output_path=output_path, **input_subswaths, dry_run=self.dry_run
         )
 
-        logger.info(f"MERGING image {self.slc_pair.slave.date} completed.")
+        if not self.dry_run:
+            logger.info(f"MERGING image {self.slc_pair.slave.date} completed.")
 
         if self.dry_run:
-            logger.debug("DRY-RUN: Creating dummy folders/files for testing purpose.")
+            logger.debug("DRY RUN: Creating dummy folders/files for testing purpose.")
             os.makedirs(output_path + ".data", exist_ok=True)
             open(output_path + ".dim", "w").close()
             pol = self.polarization.upper()
             master_datestr = format_date(self.slc_pair.master.date)
-            for channel in ("i", "q"):  # in-phase & quadrature channels
-                for suffix in ("img", "hdr"):
-                    master_filename = f"{channel}_{pol}_mst_{master_datestr}.{suffix}"
-                    open(
-                        os.path.join(output_path + ".data", master_filename), "w"
-                    ).close()
+            slave_datestr = format_date(self.slc_pair.slave.date)
+            for channel, suffix in product(("i", "q"), ("img", "hdr")):
+                master_filename = f"{channel}_{pol}_mst_{master_datestr}.{suffix}"
+                slave_filename = f"{channel}_{pol}_slv_{slave_datestr}.{suffix}"
+                for fn in (master_filename, slave_filename):
+                    open(os.path.join(output_path + ".data", fn), "w").close()
 
         return self
 
     def _finalize(self):
         self._link_master()  # make a simlink to master date
-        self._prune()  # prune unnecessary files
+        if self.prune:
+            self._prune()  # prune unnecessary files
         return self
 
     def _link_master(self):
@@ -164,22 +203,25 @@ class Sentinel1Coregistration(Coregistration):
 
         pol = self.polarization.upper()
         master_datestr = format_date(self.slc_pair.master.date)
-        for channel in ("i", "q"):  # in-phase & quadrature channels
-            for suffix in ("img", "hdr"):
-                master_filename = f"{channel}_{pol}_mst_{master_datestr}.{suffix}"
-                dst_file = os.path.join(dst_dir, "merged.data", master_filename)
-                src_file = os.path.join(src_dir, "merged.data", master_filename)
-                # copy if not exists
-                if os.path.exists(dst_file):
-                    logger.debug(f"{dst_file} already exists.")
-                    if suffix == "img":  # remove img file if exists
-                        os.remove(src_file)
-                else:
-                    logger.debug(f"Linking {src_file} to {dst_file}.")
+        for channel, suffix in product(("i", "q"), ("img", "hdr")):  # in-phase & quadrature channels
+            master_filename = f"{channel}_{pol}_mst_{master_datestr}.{suffix}"
+            dst_file = os.path.join(dst_dir, "merged.data", master_filename)
+            src_file = os.path.join(src_dir, "merged.data", master_filename)
+            # copy if not exists
+            if os.path.exists(dst_file):
+                logger.debug(f"{dst_file} already exists.")
+                if suffix == "img" and self.prune:  # remove img file if exists
+                    os.remove(src_file)
+            else:
+                logger.debug(f"Linking {src_file} to {dst_file}.")
+                if self.prune:
                     shutil.move(src_file, dst_file)
+                else:
+                    shutil.copy2(src_file, dst_file)
 
         master_symlink = os.path.join(self.output_dir, COREG_DIR, "master")
-        logger.info(f"Soft linking {self.slc_pair.master.date} to 'master' folder.")
+        if not self.dry_run:
+            logger.info(f"SOFT-LINKING {self.slc_pair.master.date} to 'master' folder.")
         if not os.path.exists(master_symlink):
             # make a symlinks to "master" folder
             os.symlink(dst_dir, master_symlink)
@@ -191,6 +233,16 @@ class Sentinel1Coregistration(Coregistration):
             # remove subswaths files (only useful before merging)
             os.remove(path + ".dim")  # type: ignore
             shutil.rmtree(path + ".data")  # type: ignore
+
+        # prune DEM folder if exists: remove all "coh" and "ifg" files
+        dem_path = os.path.join(self.output_dir, COREG_DIR, "DEM")
+        if not os.path.exists(dem_path):
+            return self
+        for root, dirs, files in os.walk(os.path.join(dem_path, "merged.data")):
+            for file in files:
+                if "ifg" in file or "coh" in file:
+                    os.remove(os.path.join(root, file))  # type: ignore
+
         return self
 
     def _serialize(self):
@@ -237,3 +289,4 @@ class TSXCoregistration(Coregistration):
 
     def coregistration(self):
         pass
+
